backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
pydantic-settings==2.5.2
SQLAlchemy==2.0.36
PyMySQL==1.1.1
python-dotenv==1.0.1
google-generativeai==0.7.2